rakuten-browser-agent/requirements.txt
browser-use
langchain-openai
google-cloud-firestore
python-dotenv
requests
playwright
